+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Changed
 - Use libpopt for command line parsing.
 - Don't start host interface task by default.
+- Separate RosInterface from rtmain.
 
 ## [1.0] - 2015-04-13
 Master thesis version